Main: Separated pipeline tools from main
