+++ pisek/__main__.py
@@ -23,12 +23,13 @@ import sys
 
 from pisek.task_config import TaskConfig
 from pisek.jobs.task_pipeline import TaskPipeline
-from pisek.jobs.server_pipelines import ServerGenKasiopea, ServerSolveKasiopea
+from pisek.server import ServerGenKasiopea, ServerSolve
 from pisek.env import Env
 from pisek.jobs.cache import Cache
 from pisek.terminal import tab, colored
+from pisek.pipeline_tools import run_pipeline, load_env
 
-from pisek import util
+from pisek.util import eprint, clean_task_dir
 from pisek.license import license, license_gnu
 import pisek.cms as cms
 from pisek.update_config import update
@@ -36,10 +37,6 @@ from pisek.visualize import visualize_command
 from pisek.extract import extract
 
 PATH = '.'
-
-def eprint(msg, *args, **kwargs):
-    print(msg, *args, file=sys.stderr, **kwargs)
-
 
 def test_task(args, **kwargs):
     return (test_task_path(PATH, **vars(args), **kwargs))
@@ -80,62 +77,10 @@ def extract_task(args, **kwargs):
         return 1
     return extract(env)
 
-
-def run_pipeline(path, pipeline, **env_args):
-    env = load_env(path, **env_args)
-    if env is None:
-        return 1
-    pipeline = pipeline(env.fork())
-    return pipeline.run_jobs(Cache(env), env)
-
-def load_env(
-        path,
-        subcommand: Optional[str] = None,
-        target: Optional[str] = None,
-        solution: Optional[str] = None,
-        full: bool = False,
-        plain: bool = False,
-        strict: bool = False,
-        clean: bool = False,
-        no_checker: bool = False,
-        solutions: Optional[list[str]] = None,
-        timeout: Optional[float] = None,
-        inputs: int = 5,
-    ):
-    config = TaskConfig()
-    err = config.load(path)
-    if err:
-        eprint(f"Error when loading config:\n{tab(err)}")
-        return None
-
-    if solutions is None:
-        solutions = config.solutions.keys()
-
-    env = Env(
-        task_dir=path,
-        config=config,
-        full=full,
-        plain=plain,
-        strict=strict,
-        no_checker=no_checker,
-        solutions=solutions,
-        timeout=timeout,
-        inputs=inputs,
-    )
-
-    if config.check_todos():
-        if env.strict:
-            eprint(colored("Unsolved TODOs in config.", env, 'red'))
-            return None
-        else:
-            eprint(colored("Warning: Unsolved TODOs in config", env, 'yellow'))
-
-    return env
-
 def clean_directory(args):
     task_dir = os.getcwd()
     eprint(f"Cleaning repository: {task_dir}")
-    util.clean_task_dir(task_dir)
+    clean_task_dir(task_dir)
 
 
 def main(argv):
