+++ pisek/util.py
@@ -30,6 +30,10 @@ DEFAULT_TIMEOUT = 360
 BUILD_DIR = "build/"
 
 
+def eprint(msg, *args, **kwargs):
+    print(msg, *args, file=sys.stderr, **kwargs)
+
+
 def rm_f(fn):
     try:
         os.unlink(fn)
